# mini-munch-backend/routes/orderRoutes.ts
import { Router } from "express";
import {
  addOrderItemController,
  createOrderController,
  deleteOrderController,
  getOrderByTableController,
  getOrdersController,
  removeOrderItemController,
  updateOrderItemController,
  updateOrderStatusController,
} from "../controllers/orderController";

// Express route handler; registers the table-related routes and maps them to their respective controllers
const router = Router();

/**
 * @openapi
 * /api/orders:
 *   post:
 *     summary: Create a DRAFT status order
 *     description: Creates an empty DRAFT status order for an occupied table. If successful, returns the created order document (consisting of the table reference, order items, total, preset DRAFT status and creation timestamp).
 *     tags:
 *       - Orders
 *     requestBody:
 *       required: true
 *       content:
 *         application/json:
 *           schema:
 *             type: object
 *             required:
 *               - tableNumber
 *             properties:
 *               tableNumber:
 *                 type: integer
 *                 minimum: 1
 *                 maximum: 20
 *     responses:
 *       201:
 *         description: Draft order was created successfully.
 *         content:
 *           application/json:
 *             schema:
 *               type: object
 *               required:
 *                 - order
 *               properties:
 *                 order:
 *                   $ref: '#/components/schemas/Order'
 *       400:
 *         description: Invalid table number.
 *       404:
 *         description: Table does not exist.
 *       409:
 *         description: Table is not occupied or already has an order.
 *       500:
 *         description: Unexpected server error.
 */
router.post("/", createOrderController);

/**
 * @openapi
 * /api/orders/table/{tableNumber}:
 *   get:
 *     summary: Get the order for a table
 *     description: Retrieves the current order document for a table, which includes the table reference, order items, total, status and creation timestamp. Each table can have at most one order while it is occupied.
 *     tags:
 *       - Orders
 *     parameters:
 *       - in: path
 *         name: tableNumber
 *         required: true
 *         schema:
 *           type: integer
 *           minimum: 1
 *           maximum: 20
 *     responses:
 *       200:
 *         description: The table order was retrieved successfully.
 *         content:
 *           application/json:
 *             schema:
 *               type: object
 *               required:
 *                 - order
 *               properties:
 *                 order:
 *                   $ref: '#/components/schemas/Order'
 *       400:
 *         description: Invalid table number.
 *       404:
 *         description: Table or order does not exist.
 *       500:
 *         description: Unexpected server error.
 */
router.get("/table/:tableNumber", getOrderByTableController);

/**
 * @openapi
 * /api/orders/{id}/items:
 *   post:
 *     summary: Add an order item to a DRAFT status order
 *     description: Adds an existing menu item to the order and stores its current name and price as a snapshot. If successful, returns the updated order document (which includes the table reference, order items, total, status and creation timestamp) with the added menu item included. Deleted menu items cannot be added to new orders.
 *     tags:
 *       - Orders
 *     parameters:
 *       - in: path
 *         name: id
 *         required: true
 *         schema:
 *           type: string
 *     requestBody:
 *       required: true
 *       content:
 *         application/json:
 *           schema:
 *             type: object
 *             required:
 *               - menuItemId
 *             properties:
 *               menuItemId:
 *                 type: string
 *     responses:
 *       200:
 *         description: Order item was added successfully.
 *         content:
 *           application/json:
 *             schema:
 *               type: object
 *               required:
 *                 - order
 *               properties:
 *                 order:
 *                   $ref: '#/components/schemas/Order'
 *       400:
 *         description: Invalid order ID or menu item ID.
 *       404:
 *         description: Order or menu item does not exist.
 *       409:
 *         description: Order is not in DRAFT status or the item is already in the order.
 *       500:
 *         description: Unexpected server error.
 */
router.post("/:id/items", addOrderItemController);

/**
 * @openapi
 * /api/orders/{id}/items/{menuItemId}:
 *   patch:
 *     summary: Change an order item quantity
 *     description: Changes the quantity of an existing order item in a DRAFT order. If successful, returns the updated order document (which includes the table reference, order items, total, status and creation timestamp) with the updated order item quality. The item can remain in the order even if it has been removed from the menu.
 *     tags:
 *       - Orders
 *     parameters:
 *       - in: path
 *         name: id
 *         required: true
 *         schema:
 *           type: string
 *       - in: path
 *         name: menuItemId
 *         required: true
 *         schema:
 *           type: string
 *     requestBody:
 *       required: true
 *       content:
 *         application/json:
 *           schema:
 *             type: object
 *             required:
 *               - quantity
 *             properties:
 *               quantity:
 *                 type: integer
 *                 minimum: 1
 *                 maximum: 10
 *     responses:
 *       200:
 *         description: Order item quantity was updated successfully.
 *         content:
 *           application/json:
 *             schema:
 *               type: object
 *               required:
 *                 - order
 *               properties:
 *                 order:
 *                   $ref: '#/components/schemas/Order'
 *       400:
 *         description: Invalid order quantity.
 *       404:
 *         description: Order or order item does not exist.
 *       409:
 *         description: Order is not in DRAFT status.
 *       500:
 *         description: Unexpected server error.
 */
router.patch("/:id/items/:menuItemId", updateOrderItemController);

/**
 * @openapi
 * /api/orders/{id}/items/{menuItemId}:
 *   delete:
 *     summary: Remove an order item from a DRAFT status order
 *     description: Removes an order item by menu item ID (even if it has been removed from the menu). If successful, returns a success response with no content.
 *     tags:
 *       - Orders
 *     parameters:
 *       - in: path
 *         name: id
 *         required: true
 *         schema:
 *           type: string
 *       - in: path
 *         name: menuItemId
 *         required: true
 *         schema:
 *           type: string
 *     responses:
 *       204:
 *         description: Item was removed successfully.
 *       400:
 *         description: Invalid order ID or menu item ID.
 *       404:
 *         description: Order or order item does not exist.
 *       409:
 *         description: Order is not in DRAFT status.
 *       500:
 *         description: Unexpected server error.
 */
router.delete("/:id/items/:menuItemId", removeOrderItemController);

/**
 * @openapi
 * /api/orders/{id}/status:
 *   patch:
 *     summary: Update an order status
 *     description: Updates an order through the valid workflow transitions DRAFT to SUBMITTED, SUBMITTED to PREPARING, and PREPARING to READY. If successful, returns the updated order document (which includes the table reference, order items, total, status and creation timestamp) with the updated order status.
 *     tags:
 *       - Orders
 *     parameters:
 *       - in: path
 *         name: id
 *         required: true
 *         schema:
 *           type: string
 *     requestBody:
 *       required: true
 *       content:
 *         application/json:
 *           schema:
 *             type: object
 *             required:
 *               - status
 *             properties:
 *               status:
 *                 type: string
 *                 enum: [SUBMITTED, PREPARING, READY]
 *     responses:
 *       200:
 *         description: Order status was updated successfully.
 *         content:
 *           application/json:
 *             schema:
 *               type: object
 *               required:
 *                 - order
 *               properties:
 *                 order:
 *                   $ref: '#/components/schemas/Order'
 *       400:
 *         description: Invalid order data, including an invalid order ID, status, or an order with no items.
 *       404:
 *         description: Order does not exist.
 *       409:
 *         description: Invalid status transition.
 *       500:
 *         description: Unexpected server error.
 */
router.patch("/:id/status", updateOrderStatusController);

/**
 * @openapi
 * /api/orders:
 *   get:
 *     summary: Get all orders for the staff dashboard
 *     description: If successful, returns all SUBMITTED, PREPARING, and READY orders for the staff dashboard (an array of order documents, each including the table reference, order items, total, status and creation timestamp), or an empty array if no such orders exist.
 *     tags:
 *       - Orders
 *     responses:
 *       200:
 *         description: Orders were retrieved successfully.
 *         content:
 *           application/json:
 *             schema:
 *               type: object
 *               required:
 *                 - orders
 *               properties:
 *                 orders:
 *                   type: array
 *                   items:
 *                     $ref: '#/components/schemas/Order'
 *       500:
 *         description: Unexpected server error.
 */
router.get("/", getOrdersController);

/**
 * @openapi
 * /api/orders/{id}:
 *   delete:
 *     summary: Delete an order
 *     description: Deletes a READY status order. If successful, returns a success response with no content.
 *     tags:
 *       - Orders
 *     parameters:
 *       - in: path
 *         name: id
 *         required: true
 *         schema:
 *           type: string
 *     responses:
 *       204:
 *         description: Order was deleted successfully.
 *       400:
 *         description: Invalid order ID.
 *       404:
 *         description: Order does not exist.
 *       409:
 *         description: Order status is not READY.
 *       500:
 *         description: Unexpected server error.
 */
router.delete("/:id", deleteOrderController);

export default router;