# mini-munch-frontend/src/services/generated/generatedApi.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * MiniMunch API
 * API and Schema for the MiniMunch restaurant ordering system!
 * OpenAPI spec version: 1.0.0
 */
export interface Table {
  _id: string;
  /**
     * @minimum 1
     * @maximum 20
     */
  tableNumber: number;
  available: boolean;
}

export interface OrderTable {
  _id: string;
  /**
     * @minimum 1
     * @maximum 20
     */
  tableNumber: number;
}

export interface MenuItem {
  _id: string;
  /**
     * @minLength 2
     * @maxLength 50
     */
  name: string;
  /**
     * @minLength 5
     * @maxLength 150
     */
  description: string;
  /** @minimum 0 */
  price: number;
}

export interface OrderItem {
  menuItem: string;
  /**
     * @minLength 2
     * @maxLength 50
     */
  name: string;
  /** @minimum 0 */
  price: number;
  /**
     * @minimum 1
     * @maximum 10
     */
  quantity: number;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  DRAFT: 'DRAFT',
  SUBMITTED: 'SUBMITTED',
  PREPARING: 'PREPARING',
  READY: 'READY',
} as const;

export interface Order {
  _id: string;
  table: OrderTable;
  items: OrderItem[];
  /** @minimum 0 */
  total: number;
  status: OrderStatus;
  createdAt: string;
}

export type GetApiTablesTableNumber200 = {
  table: Table;
};

export type PatchApiTablesTableNumberOccupy200 = {
  table: Table;
};

export type PatchApiTablesTableNumberRelease200 = {
  table: Table;
};

export type PostApiOrdersBody = {
  /**
     * @minimum 1
     * @maximum 20
     */
  tableNumber: number;
};

export type PostApiOrders201 = {
  order: Order;
};

export type GetApiOrders200 = {
  orders: Order[];
};

export type GetApiOrdersTableTableNumber200 = {
  order: Order;
};

export type PostApiOrdersIdItemsBody = {
  menuItemId: string;
};

export type PostApiOrdersIdItems200 = {
  order: Order;
};

export type PatchApiOrdersIdItemsMenuItemIdBody = {
  /**
     * @minimum 1
     * @maximum 10
     */
  quantity: number;
};

export type PatchApiOrdersIdItemsMenuItemId200 = {
  order: Order;
};

export type PatchApiOrdersIdStatusBodyStatus = typeof PatchApiOrdersIdStatusBodyStatus[keyof typeof PatchApiOrdersIdStatusBodyStatus];


export const PatchApiOrdersIdStatusBodyStatus = {
  SUBMITTED: 'SUBMITTED',
  PREPARING: 'PREPARING',
  READY: 'READY',
} as const;

export type PatchApiOrdersIdStatusBody = {
  status: PatchApiOrdersIdStatusBodyStatus;
};

export type PatchApiOrdersIdStatus200 = {
  order: Order;
};

export type GetApiMenu200 = {
  menuItems: MenuItem[];
};

export type PostApiMenuBody = {
  /**
     * @minLength 2
     * @maxLength 50
     */
  name: string;
  /**
     * @minLength 5
     * @maxLength 150
     */
  description: string;
  /** @minimum 0 */
  price: number;
};

export type PostApiMenu201 = {
  menuItem: MenuItem;
};

export type GetApiMenuId200 = {
  menuItem: MenuItem;
};